Nav links, typos, semantic subtitles
Changed nav links from <a><li></li></a> to <li><a></a></li> as only <li>
elements are allowed as a direct child of <ul>.

Corrected some minor typos, specifically "hreF" and such.

Changed subtitles from being <h2> elements, which are supposed to be
titles of a topic relating to the upper-level topic, to <p
class="subtitle"> elements, which are more correct in terms of screen
readers.